server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy>=2.0
psycopg2-binary
pgvector
pydantic
pydantic-settings
python-dotenv
openai
pdfplumber
python-docx
numpy
